+++ spec/lib/aixm/document_spec.rb
@@ -590,7 +590,7 @@ describe AIXM::Document do
               <codeType>RADIO</codeType>
               <noSeq>1</noSeq>
             </AhaUid>
-            <txtAddress>123.35 mhz</txtAddress>
+            <txtAddress>123.35</txtAddress>
             <txtRmk>A/A (callsign PUJAUT)</txtRmk>
           </Aha>
           <Sah>
@@ -1531,7 +1531,7 @@ describe AIXM::Document do
               <codeType>RADIO</codeType>
               <noSeq>1</noSeq>
             </AhaUid>
-            <txtAddress>123.35 mhz</txtAddress>
+            <txtAddress>123.35</txtAddress>
             <txtRmk>A/A (callsign PUJAUT)</txtRmk>
           </Aha>
           <Sah>
@@ -2555,7 +2555,7 @@ describe AIXM::Document do
               <codeType>RADIO</codeType>
               <noSeq>1</noSeq>
             </AhaUid>
-            <txtAddress>123.35 mhz</txtAddress>
+            <txtAddress>123.35</txtAddress>
             <txtRmk>A/A (callsign PUJAUT)</txtRmk>
           </Aha>
           <Sah>